refactor Jason logic for member messaging and moved it into Team member card and UserProfile while also making it a reusable function
