--- src/screens/Landing/members.json
@@ -1,54 +1,0 @@
-{
-  "engineers": [
-    {
-      "_id": "64f57219225338c3f72c5c34",
-      "email": "Vern76@yahoo.com",
-      "firstName": "Vern",
-      "lastName": "Greenholt",
-      "profilePicture": "https://loremflickr.com/640/480/people?lock=40866",
-      "role": "Software Engineer"
-    },
-    {
-      "_id": "64f5721a225338c3f72c5c3f",
-      "email": "Doyle_Jacobson67@yahoo.com",
-      "firstName": "Doyle",
-      "lastName": "Jacobson",
-      "profilePicture": "https://loremflickr.com/640/480/people?lock=49178",
-      "role": "Software Engineer"
-    },
-    {
-      "_id": "64f5721a225338c3f72c5c40",
-      "email": "Savion_Paucek71@yahoo.com",
-      "firstName": "Savion",
-      "lastName": "Paucek",
-      "profilePicture": "https://loremflickr.com/640/480/people?lock=69156",
-      "role": "Software Engineer"
-    }
-  ],
-  "designers": [
-    {
-      "_id": "64f57217225338c3f72c5c2a",
-      "email": "Blanca_Nicolas@yahoo.com",
-      "firstName": "Blanca",
-      "lastName": "Nicolas",
-      "profilePicture": "https://loremflickr.com/640/480/people?lock=41452",
-      "role": "UX Designer"
-    },
-    {
-      "_id": "64f57224225338c3f72c5c46",
-      "email": "star@struck.com",
-      "firstName": "Star",
-      "lastName": "Struck",
-      "profilePicture": "https://loremflickr.com/640/480/people?lock=55804",
-      "role": "UX Designer"
-    },
-    {
-      "_id": "64f57217225338c3f72c5c2b",
-      "email": "Andre8@gmail.com",
-      "firstName": "Andre",
-      "lastName": "Streich",
-      "profilePicture": "https://loremflickr.com/640/480/people?lock=18294",
-      "role": "UX Designer"
-    }
-  ]
-}
